% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ES}
\alias{ES}
\title{Functions that compute Enrichmnet Score}
\usage{
ES(stat, geneOrd, Pmiss, Nr, poz)
}
\arguments{
\item{stat}{NumericVector}

\item{geneOrd}{intiger}

\item{Pmiss}{double}

\item{Nr}{double}

\item{poz}{NumericVector}
}
\description{
Functions that compute Enrichmnet Score
}
