% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exprData}
\alias{exprData}
\title{Expression dataset.}
\format{\code{\link{data.frame}} object}
\description{
promoters of hg19 assembly of human genome on chr21 and chr22. Promoter set is
derived from refseq TSS.
}
