% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaK.R
\docType{methods}
\name{gseaK}
\alias{gseaK}
\alias{gseaK,data.frame-method}
\alias{gseaK-method}
\title{Gene Set Enrichment Analysis}
\usage{
gseaK(expr, pheno, gSets, stest = "FC", abs = TRUE, kernel = TRUE,
  n.perm = 1000, correction = TRUE, plot1 = TRUE, plot2 = TRUE,
  plot3 = TRUE)

\S4method{gseaK}{data.frame}(expr, pheno, gSets, stest = "FC", abs = TRUE,
  kernel = TRUE, n.perm = 1000, correction = TRUE, plot1 = TRUE,
  plot2 = TRUE, plot3 = TRUE)
}
\arguments{
\item{expr}{expression data}

\item{pheno}{phenotype}

\item{gSets}{Gene Sets}

\item{stest}{statistical test}

\item{abs}{absolute value}

\item{kernel}{Kernal density funcion}

\item{n.perm}{number of permutations}

\item{correction}{multiple hypothesis testing}

\item{plot1}{plot1}

\item{plot2}{plot2}

\item{plot3}{plo3}
}
\value{
returns a \code{matrix} object
}
\description{
GSEA
}
