% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{NES}
\alias{NES}
\title{Function to compute NES}
\usage{
NES(x, mp, mn)
}
\arguments{
\item{x}{ES_p vector}

\item{mp}{mean value of ES_p>=0}

\item{mn}{mean value of ES_p<0}
}
\description{
Function to compute NES
}
