% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDis}
\alias{plotDis}
\title{plot Phit+Pmiss}
\usage{
plotDis(ES_matrix, name_gene_set)
}
\arguments{
\item{ES_matrix}{\code{matrix} that stores Phit, Pmiss, and ES distributions}

\item{name_gene_set}{GS name}
}
\value{
Phit and Pmiss distributions.
}
\description{
plot Phit+Pmiss
}
