% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotES}
\alias{plotES}
\title{plot ES}
\usage{
plotES(ES_matrix, ES_obs, x_ES, pos, name_gene_set)
}
\arguments{
\item{ES_matrix}{\code{matrix} that stores ES distribution}

\item{ES_obs}{ES value}

\item{x_ES}{position of ES value}

\item{pos}{positions of genes from GS}

\item{name_gene_set}{GS name}
}
\value{
plot of ES distribution
}
\description{
plot ES
}
