% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotPerm}
\alias{plotPerm}
\title{plot of ES_null}
\usage{
plotPerm(ES_p, ES_obs, dens)
}
\arguments{
\item{ES_p}{\code{vector} that stores ES_null values}

\item{ES_obs}{ES value}

\item{dens}{Kernel density estimation}
}
\value{
plot of ES_null
}
\description{
plot of ES_null
}
