% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSet.R
\docType{methods}
\name{GSet}
\alias{GSet}
\alias{GSet,vector,character,vector,numeric,matrix,factor,matrix-method}
\alias{GSet-method}
\title{Gene Set}
\usage{
GSet(gSets, gene_name_ord, ord, n_perm, expr, class, index_ph, stest, abs)

\S4method{GSet}{vector,character,vector,numeric,matrix,factor,matrix}(gSets,
  gene_name_ord, ord, n_perm, expr, class, index_ph, stest, abs)
}
\arguments{
\item{gSets}{gene Set}

\item{gene_name_ord}{ordered gene names}

\item{ord}{ordered values of statistic}

\item{n_perm}{number of permutations}

\item{expr}{expression dataset}

\item{class}{phenotype data}

\item{index_ph}{shuffled phenotype labels}

\item{stest}{Rank metrics}

\item{abs}{absolute value}
}
\value{
returns a \code{vector} object
}
\description{
GSet
}
