% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotDis}
\alias{plotDis}
\title{plot Phit+Pmiss}
\usage{
plotDis(ES_matrix)
}
\arguments{
\item{ES_matrix}{\code{matrix} that stores Phit, Pmiss, and ES distributions}
}
\value{
Phit and Pmiss distributions.
}
\description{
plot Phit+Pmiss
}
