% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotES}
\alias{plotES}
\title{plot ES}
\usage{
plotES(ES_matrix, ES_obs, x_ES, pos)
}
\arguments{
\item{ES_matrix}{\code{matrix} that stores ES distribution}
}
\value{
Phit and Pmiss distributions.
}
\description{
plot ES
}
